chore: Moves `SynthesizeOptions` to `redteam/types` to resolve circular dependencies
